Formative Assignment for Module #2, Lesson #3: Classes
The overall goal of this assignment is to implement a Ruby class
with - class attributes, class methods, initializer methods, instance
attributes, instance methods.

The functional goal of the assignment is to write a Person class
that will maintain state thru instance and class attributes and
provide access thru instance and class class methods.